+++ package.json
@@ -21,6 +21,7 @@
     "dayjs": "^1.11.10",
     "react": "^18.2.0",
     "react-calendar": "^4.8.0",
+    "react-datepicker": "^6.1.0",
     "react-day-picker": "^8.10.0",
     "react-dom": "^18.2.0",
     "react-icons": "^5.0.1",
